Process tempo with for a few beats before processing the rest of the song
aubio tempo starts out inaccurate but gets better as the song goes on. Process the tempo without processing the pitch or onset for an arbitrary amount of beats (8 currently) before resetting the audio file and processing the whole song normally.
This makes the assumption the song does not change time signature or tempo throughout, which is a fair enough assumption.
In the future, don't use a constant number of beats to preprocess as it may not work for songs with pickups.