Round beats to the nearest measure
Instead of a hard cutoff, a note is added to the nearest measure in accordance to its duration compared to the beat determined by aubio.

Recorder has a pointer to the sheet music now. It determines when to switch to a new measure instead of the parser. The parser just tells the recorder when a certain number of beats is up.